third_party/blink/web_tests/platform/mac-mac10.10/http/tests/devtools/console/console-log-wrapped-in-framework-expected.txt

CONSOLE MESSAGE: line 13: direct console.log()
CONSOLE MESSAGE: line 7: framework log
Tests console.log() anchor location when the skip-stack-frames feature is enabled.

console-log-wrapped-in-framework.js:13 direct console.log()
console-log-wrapped-in-framework.js:14 framework log
